Add files via upload
The "Hello, World" program is the first program is the first program you learn to write in many languages, but it doesn't involve any input.

So create a program that prompts for your name and prints a greeting using your name.

Constraint:

Keep the input, string concatenation, and output separate
Display different greetings for different people. - For this item, the program discerns between names that begin with vowels and those that begin with consonants.